webParent/FontEnd/src/main/java/com/example/frontend/category/CategoryRepository.java
package com.example.frontend.category;

import com.onlineStoreCom.entity.category.Category;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;

import java.util.List;

public interface CategoryRepository extends JpaRepository<Category, Long> {


    @Query("SELECT c FROM Category c WHERE c.enable = true ORDER BY c.name ASC")
    public List<Category> findAllEnabled();

    @Query("SELECT c FROM Category c WHERE c.enable = true AND c.alias = ?1")
    public Category findByAliasEnabled(String alias);



}
